# [Nim] avoid copy for related