It was removed before, but different cc versions has different results