@jinyus 2023-10-22
Main branch:
```
Nim | 25.16 ms | 353.17 ms | 3.09 s
```

This PR:

```
Nim:

        Benchmark 1: ./related
        Processing time (w/o IO): 25.277ms
        Processing time (w/o IO): 24.997ms
        Processing time (w/o IO): 25.054ms
        Processing time (w/o IO): 25.109ms
        Processing time (w/o IO): 25.215ms
        Processing time (w/o IO): 25.445ms
        Processing time (w/o IO): 25.181ms
        Processing time (w/o IO): 25.287ms
        Processing time (w/o IO): 25.309ms
        Processing time (w/o IO): 25.229ms
        Processing time (w/o IO): 25.108ms
        Processing time (w/o IO): 25.12ms
        Processing time (w/o IO): 25.194ms
          Time (mean ± σ):      51.9 ms ±   0.7 ms    [User: 39.9 ms, System: 12.0 ms]
          Range (min … max):    50.7 ms …  53.0 ms    10 runs
         
Nim:

        Benchmark 1: ./related
        Processing time (w/o IO): 349.96ms
        Processing time (w/o IO): 351.766ms
        Processing time (w/o IO): 351.149ms
          Time (mean ± σ):     458.5 ms ±   0.3 ms    [User: 423.0 ms, System: 35.5 ms]
          Range (min … max):   458.3 ms … 458.8 ms    2 runs
         
Nim:

        Benchmark 1: ./related
        Processing time (w/o IO): 3073.467ms
        Processing time (w/o IO): 3058.321ms
        Processing time (w/o IO): 3073.271ms
          Time (mean ± σ):      3.414 s ±  0.022 s    [User: 3.317 s, System: 0.089 s]
          Range (min … max):    3.398 s …  3.430 s    2 runs
```